combine describable_indexer into descriptive_metadata_indexer
